Make ids unsigned bigints to combat constant auto increments of insert ignore

+++ database.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `videos` (
-    `id` MEDIUMINT UNSIGNED NOT NULL AUTO_INCREMENT,
+    `id` BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
     `video_id` VARCHAR(30) COLLATE utf8mb4_unicode_ci NOT NULL,
     `title` TEXT COLLATE utf8mb4_unicode_ci NOT NULL,
     `description` TEXT COLLATE utf8mb4_unicode_ci DEFAULT NULL,
@@ -17,7 +17,7 @@ CREATE TABLE `videos` (
 
 
 CREATE TABLE `channels` (
-    `id` MEDIUMINT UNSIGNED NOT NULL AUTO_INCREMENT,
+    `id` BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
     `channel_id` VARCHAR(40) NOT NULL,
     `name` VARCHAR(255) COLLATE utf8mb4_unicode_ci NOT NULL,
     `thumbnail` VARCHAR(255) DEFAULT NULL,
@@ -28,8 +28,8 @@ CREATE TABLE `channels` (
 
 
 CREATE TABLE `channelVideos` (
-    `video_id` MEDIUMINT UNSIGNED NOT NULL,
-    `channel_id` MEDIUMINT UNSIGNED NOT NULL,
+    `video_id` BIGINT UNSIGNED NOT NULL,
+    `channel_id` BIGINT UNSIGNED NOT NULL,
 
     FOREIGN KEY (`video_id`) REFERENCES `videos` (`id`),
     FOREIGN KEY (`channel_id`) REFERENCES `channels` (`id`),
@@ -39,7 +39,7 @@ CREATE TABLE `channelVideos` (
 
 
 CREATE TABLE `tags` (
-    `id` MEDIUMINT UNSIGNED NOT NULL AUTO_INCREMENT,
+    `id` BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
     `tag` VARCHAR(190) COLLATE utf8mb4_bin NOT NULL,
 
     PRIMARY KEY (`id`),
@@ -51,8 +51,8 @@ CREATE TRIGGER lcase_insert BEFORE INSERT ON `tags` FOR EACH ROW SET NEW.tag = L
 
 
 CREATE TABLE `videoTags` (
-    `video_id` MEDIUMINT UNSIGNED NOT NULL,
-    `tag_id` MEDIUMINT UNSIGNED NOT NULL,
+    `video_id` BIGINT UNSIGNED NOT NULL,
+    `tag_id` BIGINT UNSIGNED NOT NULL,
 
     FOREIGN KEY (`video_id`) REFERENCES `videos` (`id`) ON DELETE CASCADE ON UPDATE CASCADE,
     FOREIGN KEY (`tag_id`) REFERENCES `tags` (`id`) ON DELETE CASCADE ON UPDATE CASCADE,
@@ -61,7 +61,7 @@ CREATE TABLE `videoTags` (
 
 
 CREATE TABLE `playlists` (
-    `id` MEDIUMINT UNSIGNED NOT NULL AUTO_INCREMENT,
+    `id` BIGINT UNSIGNED NOT NULL AUTO_INCREMENT,
     `name` TEXT COLLATE utf8mb4_unicode_ci NOT NULL,
     `playlist_id` VARCHAR(40) COLLATE utf8mb4_unicode_ci NOT NULL,
     `site` TINYINT UNSIGNED NOT NULL,
@@ -73,8 +73,8 @@ CREATE TABLE `playlists` (
 
 
 CREATE TABLE `playlistVideos` (
-    `playlist_id` MEDIUMINT UNSIGNED NOT NULL,
-    `video_id` MEDIUMINT UNSIGNED NOT NULL,
+    `playlist_id` BIGINT UNSIGNED NOT NULL,
+    `video_id` BIGINT UNSIGNED NOT NULL,
 
     FOREIGN KEY (`playlist_id`) REFERENCES `playlists` (`id`) ON DELETE CASCADE ON UPDATE CASCADE,
     FOREIGN KEY (`video_id`) REFERENCES `videos` (`id`) ON DELETE CASCADE ON UPDATE CASCADE,
